// backend/src/modules/auth/auth.types.ts
export interface LoginDto {
  email: string;
  password: string;
}

export interface RegisterDto {
  name: string;
  email: string;
  password: string;
  rol_id?: number;
}

export interface AuthResponse {
  success: boolean;
  message: string;
  data?: {
    user: {
      id: number;
      name: string;
      email: string;
      roles: string[];
    };
    token?: string;
    dashboardRoute?: string;
  };
}